api/requirements.txt
flask
uWSGI
google-cloud-firestore
mock-firestore